Rename `user.js` to `users.js`
